+++ include/mbgl/style/sources/raster_dem_source.hpp
@@ -5,19 +5,20 @@
 #include <mbgl/util/variant.hpp>
 
 namespace mbgl {
-
 namespace style {
 
 struct RasterDEMOptions {
     std::optional<Tileset::DEMEncoding> encoding = std::nullopt;
 };
 
-class RasterDEMSource : public RasterSource {
+// NOTE: Any derived class must invalidate `weakFactory` in the destructor
+class RasterDEMSource final : public RasterSource {
 public:
     RasterDEMSource(std::string id,
                     variant<std::string, Tileset> urlOrTileset,
                     uint16_t tileSize,
                     std::optional<RasterDEMOptions> options = std::nullopt);
+    ~RasterDEMSource() override;
     bool supportsLayerType(const mbgl::style::LayerTypeInfo*) const override;
 
 protected: