+++ platform/android/MapLibreAndroid/src/cpp/map_renderer.cpp
@@ -196,6 +196,11 @@ void MapRenderer::requestSnapshot(SnapshotCallback callback) {
 
 void MapRenderer::resetRenderer() {
     renderer.reset();
+#if MLN_RENDER_BACKEND_OPENGL
+    // GL requires the context managed by the java surface for cleanup
+    // Vulkan resources can be released on any thread (finalizer in this case)
+    backend.reset();
+#endif
     swapBehaviorFlush = false;
 }